@@ -13,27 +13,30 @@
     "author": "Philip Washington Sorst <philip@sorst.net> (https://sorst.net)",
     "private": true,
     "dependencies": {
-        "@angular/animations": "^5.2.5",
-        "@angular/cdk": "^5.2.1",
-        "@angular/common": "^5.2.5",
-        "@angular/compiler": "^5.2.5",
-        "@angular/core": "^5.2.5",
-        "@angular/forms": "^5.2.5",
-        "@angular/http": "^5.2.5",
-        "@angular/material": "^5.2.1",
-        "@angular/platform-browser": "^5.2.5",
-        "@angular/platform-browser-dynamic": "^5.2.5",
-        "@angular/router": "^5.2.5",
+        "@angular/animations": "^7.2.11",
+        "@angular/cdk": "^7.3.6",
+        "@angular/common": "^7.2.11",
+        "@angular/compiler": "^7.2.11",
+        "@angular/core": "^7.2.11",
+        "@angular/forms": "^7.2.11",
+        "@angular/http": "^7.2.11",
+        "@angular/material": "^7.3.6",
+        "@angular/platform-browser": "^7.2.11",
+        "@angular/platform-browser-dynamic": "^7.2.11",
+        "@angular/router": "^7.2.11",
         "core-js": "^2.5.3",
         "faker": "^4.1.0",
         "material-design-icons": "^3.0.1",
-        "rxjs": "^5.5.6",
-        "zone.js": "^0.8.20"
+        "rxjs": "^6.4.0",
+        "tslib": "^1.9.0",
+        "zone.js": "^0.8.20",
+        "hammerjs": "^2.0.8"
     },
     "devDependencies": {
-        "@angular/cli": "^1.6.8",
-        "@angular/compiler-cli": "^5.2.5",
-        "@angular/language-service": "^5.2.5",
-        "typescript": "~2.4.2"
+        "@angular-devkit/build-angular": "~0.13.0",
+        "@angular/cli": "^7.3.7",
+        "@angular/compiler-cli": "^7.2.11",
+        "@angular/language-service": "^7.2.11",
+        "typescript": "~3.2.4"
     }
 }
