@@ -1,7 +1,6 @@
 import {Injectable} from "@angular/core";
 import {Title} from "@angular/platform-browser";
-import {BehaviorSubject} from "rxjs/BehaviorSubject";
-import {Observable} from "rxjs/Observable";
+import {BehaviorSubject, Observable} from 'rxjs';
 
 @Injectable()
 export class TitleService
